Fix JSON serialization issues and add demo files
🐛 FIXES:
- ✅ Fixed numpy float32 serialization issues in coherence.py
- ✅ Fixed entropy calculation float conversion in templated.py
- ✅ Fixed perplexity calculation float conversion in density.py
- ✅ All feature extractors now return JSON-serializable data types

📊 DEMO FILES:
- ✅ Added demo_article.txt with realistic AI slop examples
- ✅ Added demo_results.json showing complete analysis output
- ✅ Demonstrates all 10 metrics working with real data
- ✅ Shows span annotations and detailed metadata

🧪 TESTING:
- ✅ JSON output now works perfectly for automated processing
- ✅ All numeric values properly serialized
- ✅ Maintains full precision for analysis results

Ready for production use with complete API compatibility

+++ sloplint/features/coherence.py
@@ -136,13 +136,13 @@ class EntityGridCoherence:
                 emb2 = embeddings[i + 1]
 
                 # Cosine similarity
-                similarity = np.dot(emb1, emb2) / (np.linalg.norm(emb1) * np.linalg.norm(emb2))
+                similarity = float(np.dot(emb1, emb2) / (np.linalg.norm(emb1) * np.linalg.norm(emb2)))
 
                 # Convert to drift (lower similarity = higher drift)
-                drift = 1.0 - similarity
+                drift = float(1.0 - similarity)
                 drifts.append(drift)
 
-            return np.mean(drifts) if drifts else 0.0
+            return float(np.mean(drifts)) if drifts else 0.0
 
         except Exception as e:
             logger.error(f"Error calculating embedding drift: {e}")
